@@ -21,8 +21,9 @@
     "flutterwave-react-v3": "^1.0.9",
     "font-awesome": "^4.7.0",
     "google-map-react": "^2.1.9",
+    "hoist-non-react-statics": "^3.3.2",
     "moment": "^2.28.0",
-    "react": "^16.13.1",
+    "react": "^17.0.2",
     "react-accessible-accordion": "^3.3.5",
     "react-apexcharts": "^1.3.9",
     "react-bar-chart": "^0.1.2",
