@@ -1,10 +1,15 @@
 import React, { Component } from "react";
-import { Switch, Route, BrowserRouter,HashRouter } from "react-router-dom";
+import {
+  Switch,
+  Route,
+  BrowserRouter,
+  HashRouter,
+  Router,
+} from "react-router-dom";
 import "./App.css";
 import ContractorDashboard from "./Components/Contractor/ContractorDashboard";
 import NewWorkOrderForm from "./Components/Contractor/New_Work_Order";
 import ContractorWorkOrder from "./Components/Contractor/Work_Order";
-import SignIn from "./Components/Sign_in/signin";
 import SignUp from "./Components/Specialist/signup";
 import Home from "./Components/Landing_page/home";
 import NewWorkOrderStep2 from "./Components/Contractor/New_Work_Order_Step2";
@@ -12,6 +17,7 @@ import NewWorkOrderStep3 from "./Components/Contractor/New_Work_Order_Step3";
 import WorkOrderDetails from "./Components/Contractor/Work_Order_Details";
 import ContractorLanding from "./Components/Landing_page/contractorlanding";
 import Learnmore from "./Components/Landing_page/learnmore";
+import Blog from "./Components/Landing_page/blog";
 import SpecialistLanding from "./Components/Landing_page/specialistlanding";
 import Notification from "./Components/Contractor/Notification";
 import Contractor_Payment_Invoice from "./Components/Contractor/Payment_Invoice";
@@ -42,10 +48,10 @@ import Admin_Notification from "./Components/Admin/Notification";
 import All_Specialist from "./Components/Admin/All_Specialist";
 import ContactUs from "./Components/Landing_page/contactUs";
 import Projects from "./Components/Landing_page/projects";
-import IndustrySegments from  "./Components/Landing_page/IndustrySegments";
+import IndustrySegments from "./Components/Landing_page/IndustrySegments";
 import Admin_Payment_Invoice from "./Components/Admin/Payment_Invoice";
 import DeployedSpecialist from "./Components/Admin/deployedSpecialist";
-import Specialistdetails from "./Components/Admin/SpecialistPersonalDetails"
+import Specialistdetails from "./Components/Admin/SpecialistPersonalDetails";
 import "react-toastify/dist/ReactToastify.css";
 import Admin_NewWorkOrderForm from "./Components/Admin/Admin_Raise_Work_Order/New_Work_Order";
 import Admin_NewWorkOrderStep2 from "./Components/Admin/Admin_Raise_Work_Order/New_Work_Order_Step2";
@@ -54,15 +60,14 @@ import Admin_Invoice_details from "./Components/Admin/Invoice_Details";
 import AdminRaiseProformaInvoice from "./Components/Admin/RaiseProformaInvoice";
 import ProformaInvoiceAccepted from "./Components/Contractor/Invoice_Accepted";
 import ScheduledPayments from "./Components/Admin/scheduledpayments";
+import SignIn from "./Components/Sign_in/signin";
+import AdminSpecialistSettings from './Components/Admin/Specailist_Settings/specialist_settings';
 
-
-
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <div>
+const App: any = ({ history }) => {
+  return (
+    <div className="App">
+      <div>
+        <Router history={history}>
           <HashRouter>
             <Switch>
               {/* <Home /> */}
@@ -89,7 +94,7 @@ class App extends Component {
               />
               <Route path="/work_order" component={NewWorkOrderForm} />
               <Route path="/specialistlanding" component={SpecialistLanding} />
-              <Route path="/Learnmore" component={Learnmore} />
+              <Route path="/blog" component={Blog} />
               <Route path="/projects" component={Projects} />
               <Route path="/contactus" component={ContactUs} />
               <Route path="/contractorlanding" component={ContractorLanding} />
@@ -107,7 +112,7 @@ class App extends Component {
                 path="/contractor_work_order_step3"
                 component={NewWorkOrderStep3}
               />
-              
+
               <Route
                 path="/contractor_work_order_details"
                 component={WorkOrderDetails}
@@ -117,15 +122,21 @@ class App extends Component {
                 path="/payment_invoice"
                 component={Contractor_Payment_Invoice}
               />
-              
-              <Route path="/admin_invoice_details/:id" component={Admin_Invoice_details} />
-              <Route path="/invoice_details/:id" component={Invoice_details} />
-              <Route path="/raise_proforma_invoice" component={AdminRaiseProformaInvoice} />
 
-              
+              <Route
+                path="/admin_invoice_details/:id"
+                component={Admin_Invoice_details}
+              />
+              <Route path="/invoice_details/:id" component={Invoice_details} />
+              <Route
+                path="/raise_proforma_invoice"
+                component={AdminRaiseProformaInvoice}
+              />
+
               <Route
                 path="/contractor_profile"
                 component={Contractor_Profile}
+                exact={true}
               />
               {/* Contractor Dashbaord */}
               {/* Admin Dashboard */}
@@ -142,7 +153,7 @@ class App extends Component {
 
               <Route path="/admin_dashboard" component={AdminDashboard} />
               <Route path="/scheduled_payments" component={ScheduledPayments} />
-              
+
               <Route path="/contractor_list" component={ListOfContractor} />
               <Route
                 path="/contractor_onboarding"
@@ -153,22 +164,25 @@ class App extends Component {
                 path="/contractor_onboarding"
                 component={ContractorOnboarding}
               />
-               <Route
+              <Route
                 path="/proforma_invoice_accepted"
                 component={ProformaInvoiceAccepted}
               />
-              
+
               <Route
                 path="/deployedspecialist"
                 component={DeployedSpecialist}
               />
-              <Route  
+              <Route
                 path="/admin_work_details"
                 component={AdminViewWorkOrderDetails}
               />
               <Route path="/allspecialist" component={All_Specialist} />
-              <Route path="/admin_payment_invoice" component={Admin_Payment_Invoice} />
-              
+              <Route
+                path="/admin_payment_invoice"
+                component={Admin_Payment_Invoice}
+              />
+
               <Route
                 path="/admin_evaluation_step4"
                 component={AdminWorkOrderEvaluationStep4}
@@ -183,18 +197,24 @@ class App extends Component {
                 component={AdminWorkOrderEvaluationStep2}
               />
               <Route path="/admin_work_order" component={AdminWorkOrder} />
-              <Route path="/admin_new_work_order" component={Admin_NewWorkOrderForm} />
-              <Route path="/admin_new_work_order_step2" component={Admin_NewWorkOrderStep2} />
-              <Route path="/admin_new_work_order_step3" component={Amin_NewWorkOrderStep3} />
-              
+              <Route
+                path="/admin_new_work_order"
+                component={Admin_NewWorkOrderForm}
+              />
+              <Route
+                path="/admin_new_work_order_step2"
+                component={Admin_NewWorkOrderStep2}
+              />
+              <Route
+                path="/admin_new_work_order_step3"
+                component={Amin_NewWorkOrderStep3}
+              />
+              <Route path="/admin/specialist/settings/:id" component={AdminSpecialistSettings} />
               <Route
                 path="/admin_assign_specialist"
                 component={AssignSpecialist}
               />
-                <Route
-                path="/specialistdetails"
-                component={Specialistdetails}
-              />
+              <Route path="/specialistdetails" component={Specialistdetails} />
 
               {/* Admin Dashboard ends */}
               {/* specialist Dashboard */}
@@ -211,24 +231,29 @@ class App extends Component {
               <Route path="/settings" component={SpecialistSettings} />
               {/* specialist Dashboard */}
               <Route path="/specialist_signup" component={SignUp} />
-              {/* <Route path="/signin" component={signIn} /> */}
               <Route
+                exact={true}
                 path="/specialistnotifications"
                 component={SpecialistNotification}
               />
               <Route
+                exact={true}
                 path="/Specialist_Payment_Invoice/:id"
                 component={Specialist_Payment_Invoice}
               />
-              <Route path="/signin" component={SignIn} />
-              <Route path="/forgot_password" component={ForgotPassword} />
-              <Route path="/" component={Home} />
+              <Route path="/signin" exact={true} component={SignIn} />
+              <Route
+                path="/forgot_password"
+                exact={true}
+                component={ForgotPassword}
+              />
+              <Route path="/" exact={true} component={Home} />
             </Switch>
           </HashRouter>
-        </div>
+        </Router>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
 export default App;
