@@ -10,8 +10,6 @@ import eye from "../../images/eye.png";
 import eyeclose from "../../images/eye-off.png";
 import NavBar from "../Widgets/navigation";
 
-
-
 const Contractorsignup = withRouter((props) => {
   const [state, setState] = useState({
     email: "",
@@ -85,18 +83,18 @@ const Contractorsignup = withRouter((props) => {
         errorMessage: "Company name is required",
       });
     }
-    if (!website_url) {
-      return setState({
-        ...state,
-        errorMessage: "website url is required",
-      });
-    }
-    if (!industry) {
-      return setState({
-        ...state,
-        errorMessage: "industry is required",
-      });
-    }
+    // if (!website_url) {
+    //   return setState({
+    //     ...state,
+    //     errorMessage: "website url is required",
+    //   });
+    // }
+    // if (!industry) {
+    //   return setState({
+    //     ...state,
+    //     errorMessage: "industry is required",
+    //   });
+    // }
     submitForm();
   };
   const submitForm = () => {
@@ -106,7 +104,7 @@ const Contractorsignup = withRouter((props) => {
     });
     const data = {
       email,
-      password,
+      // password,
       first_name,
       last_name,
       help,
@@ -121,14 +119,15 @@ const Contractorsignup = withRouter((props) => {
       .then((res) => {
         console.log(res.data);
         localStorage.setItem("loggedInDetails", JSON.stringify(res.data));
-        if (res?.data?.user_type == "contractor") {
-          return props.history.push("/contractor_dashboard");
-        }
+        // if (res?.data?.user_type == "contractor") {
+        //   return props.history.push("/contractor_dashboard");
+        // }
         setState({
           ...state,
           isloading: false,
-          successMessage: "Successfully signed up",
+          successMessage: "Successfully signed up, your account would be activated within the next 48hours",
         });
+        window.scrollTo(0,0)
       })
       .catch((err) => {
         console.log(err?.response);
@@ -175,19 +174,19 @@ const Contractorsignup = withRouter((props) => {
   const hidePassword = () => {
     setState({
       ...state,
-      passwordIsOpen: passwordIsOpen ? false : true
+      passwordIsOpen: passwordIsOpen ? false : true,
     });
   };
   const fieldRef: any = useRef();
   useEffect(() => {
-    if (errorMessage || successMessage && fieldRef) {
+    if (errorMessage || (successMessage && fieldRef)) {
       fieldRef.current.scrollIntoView({
-        behavior: "smooth"
+        behavior: "smooth",
       });
     }
   }, [errorMessage, successMessage]);
   useEffect(() => {
-    window.scrollTo(-0,-0);
+    window.scrollTo(-0, -0);
     Axios.all([
       Axios.get<any, AxiosResponse<any>>(`${API}/industries`),
       Axios.get<any, AxiosResponse<any>>(`${API}/contractor-needs`),
@@ -208,9 +207,9 @@ const Contractorsignup = withRouter((props) => {
       });
   }, []);
   console.log(listOfIndustries);
-  return (
+  return(
     <div>
-       <NavBar />
+      <NavBar />
       <section className="forms-section">
         <div className="forms-section-image"></div>
         <div className="formwrplift">
@@ -221,14 +220,11 @@ const Contractorsignup = withRouter((props) => {
                   <div className="padded-form-wrapper">
                     <div className="form-header">
                       <h4 className="form-title">Sign up to get Technical Specialist</h4>
-                      <p>Sign up to get Technical Specialists</p>
                     </div>
                     <div className="form-descr-text">
                       <p>
                       Connect with a member of our team to explore how MolecularTech can support you and your business to get technical specialists.
-                      </p>
-                      <p>
-                        Leave some information about you & company, and we’ll contact you within 24 hours.
+                      Leave some information about you & company, and we’ll contact you within 24 hours.
                       </p>
                     </div>
                   </div>
@@ -237,12 +233,16 @@ const Contractorsignup = withRouter((props) => {
                       <Col md={12} className="col_space">
                         {errorMessage && (
                           <div className="text-center">
-                            <Alert variant="danger" className="cntralertmessg">{errorMessage}</Alert>
+                            <Alert variant="danger" className="cntralertmessg">
+                              {errorMessage}
+                            </Alert>
                           </div>
                         )}
                         {successMessage && (
                           <div className="text-center ">
-                            <Alert variant="success" className="cntralertmessg">{successMessage}</Alert>
+                            <Alert variant="success" className="cntralertmessg">
+                              {successMessage}
+                            </Alert>
                           </div>
                         )}
                       </Col>
@@ -260,7 +260,7 @@ const Contractorsignup = withRouter((props) => {
                             name="first_name"
                             value={first_name}
                             onChange={onchange}
-                            placeholder="Enter your first name"
+                            // placeholder="Enter your first name"
                             size={75}
                             className="form-control forminput"
                           />
@@ -277,7 +277,7 @@ const Contractorsignup = withRouter((props) => {
                             name="last_name"
                             value={last_name}
                             onChange={onchange}
-                            placeholder="Enter your Last name"
+                            // placeholder="Enter your Last name"
                             size={75}
                             className="form-control forminput"
                           />
@@ -294,14 +294,14 @@ const Contractorsignup = withRouter((props) => {
                             name="email"
                             value={email}
                             onChange={onchange}
-                            placeholder="Enter your email"
-                            size={75}
+                            // placeholder="Enter your email"
+                            // size={75}
                             className="form-control forminput"
                           />
                         </label>
                       </Col>
                     </Row>
-                    <Row>
+                    {/* <Row>
                       <Col md={12}>
                         <label className="inputlabel">
                           <span className="rdfrmlbl">
@@ -313,30 +313,30 @@ const Contractorsignup = withRouter((props) => {
                             name="password"
                             onChange={onChangepassword}
                             value={password}
-                            placeholder="Enter your password"
-                            size={75}
+                            // placeholder="Enter your password"
+                            // size={75}
                             className="form-control forminput"
                           />
                         </label>
                         <div className="text-right">
-                  {passwordIsOpen ? (
-                    <img
-                      src={eye}
-                      className="hideeye"
-                      onClick={hidePassword}
-                      alt="hideeye"
-                    />
-                  ) : (
-                    <img
-                      src={eyeclose}
-                      className="hideeye"
-                      onClick={hidePassword}
-                      alt="hideeye"
-                    />
-                  )}
-                </div>
+                          {passwordIsOpen ? (
+                            <img
+                              src={eye}
+                              className="hideeye"
+                              onClick={hidePassword}
+                              alt="hideeye"
+                            />
+                          ) : (
+                            <img
+                              src={eyeclose}
+                              className="hideeye"
+                              onClick={hidePassword}
+                              alt="hideeye"
+                            />
+                          )}
+                        </div>
                       </Col>
-                    </Row>
+                    </Row> */}
                     <Row>
                       <Col md={12}>
                         <label className="inputlabel">
@@ -349,7 +349,7 @@ const Contractorsignup = withRouter((props) => {
                             name="company_name"
                             value={company_name}
                             onChange={onchange}
-                            placeholder="Enter your Company Name"
+                            // placeholder="Enter your Company Name"
                             size={75}
                             className="form-control forminput"
                           />
@@ -369,7 +369,7 @@ const Contractorsignup = withRouter((props) => {
                             name="phone"
                             value={phone}
                             onChange={onchange}
-                            placeholder="Enter your Phone Number"
+                            // placeholder="Enter your Phone Number"
                             size={96}
                             className="form-control forminput"
                           />
@@ -408,7 +408,7 @@ const Contractorsignup = withRouter((props) => {
                         name="website_url"
                         value={website_url}
                         onChange={onchange}
-                        placeholder="E.g http://www.example.com"
+                        // placeholder="E.g http://www.example.com"
                         size={96}
                         className="form-control forminput"
                       />
@@ -454,10 +454,14 @@ const Contractorsignup = withRouter((props) => {
                         className="form-btn form-btnactive"
                         onClick={submitForm}
                       >
-                        {!isloading ? "Create Account" : "Creating Account"}
+                        {!isloading ? "Create Account" : "Processing..."}
                       </span>
                     </div>
-                    <Link to="/signin"><p className="signuprgqt">Have Molecular account?<span>Login</span></p></Link>
+                    <Link to="/signin">
+                      <p className="signuprgqt">
+                        Have Molecular account?<span> Login</span>
+                      </p>
+                    </Link>
                   </div>
                 </form>
               </Col>
