@@ -19,8 +19,21 @@ import Footer from "./footer";
 import Work from "./workandbuild";
 import AOS from "aos";
 import "aos/dist/aos.css";
+import TextTransition, { presets } from "react-text-transition";
+import {
+  BlinkingCursorTextBuilder,
+  FloatingLettersTextBuilder,
+} from "react-animated-text-builders";
+
+const TEXTS = [
+  "Efficiently, deployed, vetted, certified pipeline specialist",
+  "Let's do the hard work",
+  "Let's empower the grey collar workers",
+];
 
 const Home = () => {
+  const [index, setIndex] = React.useState(0);
+
   useEffect(() => {
     window.scrollTo(-0, -0);
     AOS.init({
@@ -29,10 +42,35 @@ const Home = () => {
     AOS.refresh();
   }, []);
 
+  React.useEffect(() => {
+    const intervalId = setInterval(
+      () => setIndex((index) => index + 1),
+      3000 // every 3 seconds
+    );
+    return () => clearTimeout(intervalId);
+  }, []);
+  // useEffect(()=>{
+  //   setInterval(()=>{
+  //     if (current.length >30){
+  //       setState({
+  //         ...state,
+  //         current: `Let's ${" "}  do ${" "}  the ${" "}  hard ${" "}  work`
+  //       })
+  //     }
+
+  //     if (current.length <30){
+  //       setState({
+  //         ...state,
+  //         current: `Let's ${" "}  do ${" "}  the ${" "}  hard  ${" "} work`
+  //       })
+  //     }
+  //   },4000)
+  // })
   const [state, setState] = useState({
     toggleSections: true,
+    current: `Let's ${" "} build ${"  "} more  ${" "} pipelines`,
   });
-  const { toggleSections } = state;
+  const { toggleSections, current } = state;
 
   const contractorSection = () => {
     setState({
@@ -46,6 +84,7 @@ const Home = () => {
       toggleSections: false,
     });
   };
+
   return (
     <div>
       <div id="home"></div>
@@ -56,7 +95,24 @@ const Home = () => {
             <Col md={12} className="fixedti22">
               <div className="section-one-content">
                 <h4 className="section-one-heading">
-                  Let’s Build the Natural Gas Pipeline to Global Standards
+                  <TextTransition
+                    text={TEXTS[index % TEXTS.length]}
+                    springConfig={presets.wobbly}
+                    direction="up"
+                    className="trans3"
+                  />
+
+                  {/* <FloatingLettersTextBuilder
+                    floatingSpeed={500}
+                    lettersAppearanceDelay={250}
+                    animationMaxMargin={"150px"}
+                    animationMinMargin={"0px"}
+                    style={{
+                      justifyContent: "center",
+                    }}
+                  >
+                     Let's &nbsp; build&nbsp; more &nbsp; gas &nbsp; pipelines
+                  </FloatingLettersTextBuilder> */}
                 </h4>
                 {/* <p className="section-one-descriptn">
                   Get special skill work done and transforming how work gets
@@ -122,12 +178,17 @@ const Home = () => {
                 />
               </div>
               <p className="section-three-heading">
-              To empower the people who do the hard work of powering Africa through natural gas pipelines
+                To empower the people who do the hard work of powering Africa
+                through natural gas pipelines
               </p>
             </div>
           </Col>
           <Col md={6} data-aos="zoom-in-up" className="sctn3col2">
-            <img src={oilengineers} className="img-fluid oilengimg" alt="pic3" />
+            <img
+              src={oilengineers}
+              className="img-fluid oilengimg"
+              alt="pic3"
+            />
           </Col>
         </Row>
       </div>
@@ -166,58 +227,104 @@ const Home = () => {
                 them
               </p>
             </div>
-            {/* <div className="homepg-cards">
+            <div className="homepg-cards">
               <div className="homepg-cards-title">
                 <img src={greencheck} alt="pic7" />
-                <p>Certify Specialists</p>
+                <p>Convenience and Speed</p>
+              </div>
+              <p></p>
+            </div>
+          </div>
+          {/* second */}
+          <div className="homepg-cards-wrapper">
+            <div className="homepg-cards">
+              <div className="homepg-cards-title">
+                <img src={star} alt="pic4" />
+                <p>More Work Opportunities</p>
               </div>
               <p>
-                We connect you with the workers you need. No matter what you're
-                looking for, our diverse workforce can meet your unique needs.
+                We’ve built the most robust workforce marketplace for technical
+                specialists in Africa. We’re always looking for your next work
+                opportunities from several EPC Contractors
               </p>
-            </div> */}
+            </div>
+            <div className="homepg-cards">
+              <div className="homepg-cards-title">
+                <img src={target} alt="pic5" />
+                <p>Better Work Conditions</p>
+              </div>
+              <p>
+                Our dedicated recruiting team is committed to identifying the
+                right opportunities aligned with your unique preferences and
+                goals.
+              </p>
+            </div>
+            <div className="homepg-cards">
+              <div className="homepg-cards-title">
+                <img src={meter} alt="pic6" />
+                <p>Better and Timely Pay</p>
+              </div>
+              <p>
+                Our technical Specialists benefit from the most flexible payment
+                terms in the industry. We ensure you get paid what you deserve
+                plus other great benefits such as Health Insurance and
+                Professional Indemnity.
+              </p>
+            </div>
+            <div className="homepg-cards">
+              <div className="homepg-cards-title">
+                <img src={greencheck} alt="pic7" />
+                <p>Continous Training and Certification</p>
+              </div>
+              <p></p>
+            </div>
           </div>
         </Container>
         <div id="our_services"></div>
       </div>
-        <div className="section-five">
-          <Row>
-            <Col md={6} className="section-five-img-col">
-              <img src={lead} className="img-fluid fluid8" alt="pic8" />
-            </Col>
-            <Col md={6} className="section-five-content">
-              <div className="section-five-subheading">
-                <span> What we do </span>
-              </div>
-              <p className="section-five-paragp1">
-                We Empower the People Who Do the Hard Work of Powering Africa
-                though Natural Gas Pipelines.
-              </p>
-              <div className="section-fivebtn-wrapp">
-                <Link to="/contractorlanding">
+      <div className="section-five">
+        <Row>
+          <Col md={6} className="section-five-img-col">
+            <img src={lead} className="img-fluid fluid8" alt="pic8" />
+          </Col>
+          <Col md={6} className="section-five-content">
+            <div className="section-five-subheading">
+              <span> What we do </span>
+            </div>
+            <div className="section-five-paragp1">
+              We Empower the People Who Do the Hard Work of Powering Africa
+              though Natural Gas Pipelines.
+            </div>
+            <div>
+              <Link to="/learnmore" className="btn-text">
+                Learn more &rarr;
+              </Link>
+            </div>
+            <div className="section-fivebtn-wrapp">
+              <Link to="/contractorlanding">
                 <span className="home-btn-primary contrtor-btn">
                   EPC Contractors
                 </span>
-                </Link>
-                {/* <span
+              </Link>
+              {/* <span
                   onClick={specialistSection}
                   className="section-five-btn scfvspclstbtn"
                 > */}
-                <Link to="/specialistlanding">
+              <Link to="/specialistlanding">
                 <span
                   className="home-btn-primary specilst-btn"
                   onClick={specialistSection}
                 >
                   Pipeline Specialists
                 </span>
-                </Link>
-              </div>
-              {/* <p className="section-five-paragp2">
+              </Link>
+            </div>
+            {/* <p className="section-five-paragp2">
                 Access reliable workforce.
                 <br /> Our platform meets your need as a contractor, as well as
                 a specialised skilled worker.
               </p> */}
-              {/* <p className="section-five-paragp3">
+            {/* <p className="section-five-paragp3">
                 MolecularTech aggregates end-to-end natural gas pipeline
                 operations to help Engineering, Procurement & Construction(EPC)
                 Operators in the upstream, midstream and downstream sectors of
@@ -226,16 +333,16 @@ const Home = () => {
                 Steering Professionals, on the other hand, to achieve efficient
                 deployment and engagement for Pipeline Construction Projects.
               </p> */}
-              {/* <div className="section-five-btn-wrap">
+            {/* <div className="section-five-btn-wrap">
                 <Link to="/contractor_signup">
                   <span className="home-btn-primary sectnfivecontrtor-btn">
                     Hire Specialists
                   </span>
                 </Link>
               </div> */}
-            </Col>
-          </Row>
-        </div>
+          </Col>
+        </Row>
+      </div>
       {/* {toggleSections === false && (
         <div className="section-five">
           <Row>
@@ -298,8 +405,9 @@ const Home = () => {
       )} */}
       <div className="section-six">
         <h4 data-aos="fade-down">
-          We use technology to streamline everything from recruiting to payments.
-           We're focused on delivering qualityoutcomes for EPC Contractors and Technical Specialists
+          We use technology to streamline everything from recruiting to
+          payments. We're focused on delivering quality outcomes for EPC
+          Contractors and Technical Specialists
         </h4>
         {/* <p>Jennifer Ghan — CEO</p> */}
       </div>
