@@ -9,73 +9,87 @@ import eye from "../../images/eye.png";
 import eyeclose from "../../images/eye-off.png";
 import NavBar from "../Widgets/navigation";
 
-const SignUp = withRouter( (props:any) => {
- 
-const [state, setState] = useState({
-  firstName: "",
-  lastName: "",
-  email: "",
-  phone: "",
-  password: "",
-  skill:"",
-  isLoading: false,
-  errorMessage: "",
-  successMessage: "",
-  btnState: false,
-  jobs: [],
-  passwordIsOpen: true
-});
-const{
-  firstName,
-  lastName,
-  email,
-  phone,
-  password,
-  skill,
-  isLoading,
-  errorMessage,
-  successMessage,
-  btnState,
-  jobs,
-  passwordIsOpen
-} = state;
-const onSubmit = () => {
-  setState({ ...state, isLoading: true });
-  const data = {
-    first_name: firstName,
-    last_name: lastName,
-    email: email,
-    phone: phone,
-    password: password,
-    skill: skill
-  };
-  console.log(data);
-  //posting data to the api
-  axios
-    .post(`${API}/register/specialist`, data)
-    .then((response) => {
-      console.log(response);
-      if (response.status === 200) {
-        //store name and email to local storage
-        const userdata: any = [];
-        userdata.push(email,firstName);
-        localStorage.setItem("userdata", JSON.stringify(userdata));
-        //store user token to to local storage 
-        localStorage.setItem("loggedInDetails", JSON.stringify(response.data));
-        //push to otp page
-        setTimeout(() => {
-          props?.history?.push("/molecular_otp");
-          console.log(props);
-        }, 3000);
+const SignUp = withRouter((props: any) => {
+  const [state, setState] = useState({
+    firstName: "",
+    lastName: "",
+    email: "",
+    phone: "",
+    password: "",
+    skill: "",
+    isLoading: false,
+    errorMessage: "",
+    successMessage: "",
+    btnState: false,
+    jobs: [],
+    passwordIsOpen: true,
+  });
+  const {
+    firstName,
+    lastName,
+    email,
+    phone,
+    password,
+    skill,
+    isLoading,
+    errorMessage,
+    successMessage,
+    btnState,
+    jobs,
+    passwordIsOpen,
+  } = state;
+  const onSubmit = () => {
+    setState({ ...state, isLoading: true });
+    const data = {
+      first_name: firstName,
+      last_name: lastName,
+      email: email,
+      phone: phone,
+      skill: skill,
+    };
+    console.log(data);
+    //posting data to the api
+    axios
+      .post(`${API}/register/specialist`, data)
+      .then((response) => {
+        console.log(response);
+        if (response.status === 200) {
+          //store name and email to local storage
+          const userdata: any = [];
+          userdata.push(email, firstName);
+          localStorage.setItem("userdata", JSON.stringify(userdata));
+          //store user token to to local storage
+          localStorage.setItem(
+            "loggedInDetails",
+            JSON.stringify(response.data)
+          );
+          //push to otp page
+          // setTimeout(() => {
+          //   props?.history?.push("/molecular_otp");
+          //   console.log(props);
+          // }, 3000);
+          window.scrollTo(0, 0);
+          setState({
+            ...state,
+            successMessage: "Thanks for signing up on molecular platform! We are currently reviewing your application and will get back to you shortly.",
+            isLoading: false,
+          });
+        }
+      })
+      .catch((err) => {
+        console.log(err?.response)
+        if (err?.response?.status == 406) {
+          return setState({
+            ...state,
+            errorMessage: err?.response?.data?.errors?.email?.join(""),
+          });
+        }
         setState({
           ...state,
-          errorMessage: "signup failed, check your internet connection",
-          isLoading: false,
+          errorMessage: "Registration failed",
         });
-      }
       });
-  }
-  
+  };
 
   const onChangeHandler = (e) => {
     setState({
@@ -114,7 +128,7 @@ const onSubmit = () => {
   }, [errorMessage, successMessage]);
   const validateForm = (e) => {
     e.preventDefault();
-    if (firstName == "" && lastName == "" && email == "" && password == "") {
+    if (firstName == "" && lastName == "" && email == "") {
       return setState({
         ...state,
         errorMessage: "please enter your details",
@@ -144,12 +158,13 @@ const onSubmit = () => {
         errorMessage: "Please enter your skill",
       });
     }
-    if (password == "") {
-      return setState({
-        ...state,
-        errorMessage: "Please enter your password",
-      });
-    } else {
+    // if (password == "") {
+    //   return setState({
+    //     ...state,
+    //     errorMessage: "Please enter your password",
+    //   });
+    // }
+    else {
       onSubmit();
     }
   };
@@ -170,7 +185,7 @@ const onSubmit = () => {
   }, []);
   return (
     <div>
-       <NavBar />
+      <NavBar />
       <section className="forms-section">
         <div className="specialistforms-section-image"></div>
         <div className="formwrplift">
@@ -188,8 +203,10 @@ const onSubmit = () => {
                     </div>
                     <div className="form-descr-text">
                       <p>
-                      Leave some information about you in the form below, attach your resume and we’ll get in touch with you.
-                      Learn what to expect when you sign up to find work with MolecularTech
+                        Leave some information about you in the form below,
+                        attach your resume and we’ll get in touch with you.
+                        Learn what to expect when you sign up to find work with
+                        MolecularTech
                       </p>
                     </div>
                   </div>
@@ -216,7 +233,7 @@ const onSubmit = () => {
                             name="firstName"
                             value={firstName}
                             onChange={onChangeHandler}
-                            placeholder="Enter your first name"
+                            // placeholder="Enter your first name"
                             size={75}
                             className="form-control forminput"
                           />
@@ -233,7 +250,7 @@ const onSubmit = () => {
                             name="lastName"
                             value={lastName}
                             onChange={onChangeHandler}
-                            placeholder="Enter your Last name"
+                            // placeholder="Enter your Last name"
                             size={75}
                             className="form-control forminput"
                           />
@@ -250,7 +267,7 @@ const onSubmit = () => {
                         name="email"
                         value={email}
                         onChange={onChangeHandler}
-                        placeholder="Enter your Email Address"
+                        // placeholder="Enter your Email Address"
                         size={96}
                         className="form-control forminput"
                       />
@@ -265,7 +282,7 @@ const onSubmit = () => {
                         name="phone"
                         value={phone}
                         onChange={onChangeHandler}
-                        placeholder="Enter your Phone Number"
+                        // placeholder="Enter your Phone Number"
                         size={96}
                         className="form-control forminput"
                       />
@@ -299,7 +316,7 @@ const onSubmit = () => {
                         </div>
                       </Col>
                     </Row>
-                    <label className="inputlabel">
+                    {/* <label className="inputlabel">
                       <span className="rdfrmlbl">
                         Password<span className="asteric">*</span>
                       </span>
@@ -307,13 +324,13 @@ const onSubmit = () => {
                         type={passwordIsOpen ? "password" : "text"}
                         name="password"
                         value={password}
-                        placeholder="******"
-                        onChange={onChangepassword}
+                        // placeholder="******"
+                        // onChange={onChangepassword}
                         size={96}
                         className="form-control forminput"
                       />
-                    </label>
-                    <div className="text-right">
+                    </label> */}
+                    {/* <div className="text-right">
                       {passwordIsOpen ? (
                         <img
                           src={eye}
@@ -329,7 +346,7 @@ const onSubmit = () => {
                           alt="hideeye"
                         />
                       )}
-                    </div>
+                    </div> */}
                     <div className="form-check">
                       <label className="form-check-label">
                         <input
@@ -338,7 +355,7 @@ const onSubmit = () => {
                           value=""
                           id="flexCheckDefault"
                         />
-                        Creating an account means you’re okay with our Terms of
+                        Creating an account means you agree to our Terms of
                         Service, Privacy Policy, and our
                         <br />
                         default Notification Settings.
@@ -356,7 +373,7 @@ const onSubmit = () => {
                     </div>
                     <Link to="/signin">
                       <p className="signuprgqt">
-                        Have Molecular account?<span>Login</span>
+                        Have Molecular account?<span> Login</span>
                       </p>
                     </Link>
                   </div>
