@@ -1,5 +1,6 @@
 import React from 'react';
 import './navigation.css';
+import { Link } from "react-router-dom";
 import logo from '../../images/Molecular.png';
 import Burger from './burger'
 
@@ -8,7 +9,7 @@ const NavBar =()=>{
     return(
         <div>
            <div className="nav">
-                  <div className="nav-logo"><img src={logo}/></div>
+                  <Link to="/#home"><div className="nav-logo"><img src={logo}/></div></Link>
                  <Burger/>
               </div>
         </div>
