@@ -2,6 +2,8 @@ import { ToastContainer, toast } from "react-toastify";
 const moment = require("moment");
 // Molecular Production API
 export const API = "https://staging-api.molecularpro.co/api/v1";
+// const live API = "https://api.molecularpro.co/api/v1/"
+export const MID = "GP0000001"
 export const capitalize = (s) => {
   if (typeof s !== "string") return "";
   return s.charAt(0).toUpperCase() + s.slice(1);
@@ -65,7 +67,7 @@ export const getAdminToken=()=>{
     ? JSON.parse(availableToken)
     : window.location.assign("/");
   if (token.user_type !== "admin") {
-    return window.location.assign = "/login";
+    return window.location.assign = "/signin";
   }
   return token
 }
@@ -88,8 +90,13 @@ export const splitName =(x)=>{
   return Math.abs(num) > 999 ? Math.sign(num)*((Math.abs(num)/1000).toFixed(1)) + 'k' : Math.sign(num)*Math.abs(num)
 }
 export const current_currency = "₦"
-// // Clarity Production API
-// export const API = "http://api.molecular.elostage.xyz/api/v1"
-// // Clarity Staging API
-// // export const API = "https://claritydev.herokuapp.com";
-// // export const webSocketAPI = "claritydev.herokuapp.com";
+
+export const formatTime2 = (date) => {
+  if (date) {
+    const dateTime = moment(date).format('YYYY-MM-DD');
+    console.log(dateTime.currentTime)
+    return dateTime;
+  } else {
+    return "";
+  }
+};
