@@ -3,11 +3,18 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
-import { HashRouter as Switch, Route, BrowserRouter, HashRouter } from "react-router-dom";
+import {
+  HashRouter as Switch,
+  Route,
+  BrowserRouter,
+  HashRouter,
+} from "react-router-dom";
+import { createBrowserHistory } from "history";
 
+const history = createBrowserHistory();
 ReactDOM.render(
   <HashRouter>
-    <App />
+    <App history={history} />
   </HashRouter>,
   document.getElementById("root")
 );
